refactor(solver): split variance analysis out of infer_resolve
`infer_resolve.rs` was already over the 2000-line hard ceiling on main
(2026 lines); the inference fix in this branch pushed it further. Move
the self-contained variance-analysis concern (`VarianceState` plus
`compute_variance`/`compute_variance_helper`/`get_variance`) into a new
`infer_variance.rs` module. Behavior-preserving: the methods keep the
same signatures on `InferenceContext`, so no call sites change.

infer_resolve.rs drops to 1896 lines and is no longer oversized,
reducing the solver oversized-file count.
